refactor(webui): extract plugin README display into PluginReadme component
